Refactoring of code, import env, gitignore
