+++ go-backend/main.go
@@ -1,29 +1,23 @@
 package main
 
 import (
-	"Identity-Lock/go-backend/middleware"
-
+	"Identity-Lock/go-backend/api"
 	"log"
-	"net/http"
 
-	"github.com/gorilla/handlers"
-	"github.com/gorilla/mux"
+	"github.com/joho/godotenv"
 )
 
-func pong(w http.ResponseWriter, r *http.Request) {
-	w.Write([]byte("pong"))
-}
-
 func main() {
-	r := mux.NewRouter()
-	r.HandleFunc("/ping", pong)
+	// Load env variables
+	err := godotenv.Load()
 
-	r.Use(middleware.LoggingMiddleware)
-	r.Use(middleware.AuthMiddleware)
+	if err != nil {
+		log.Fatalf("Error loading .env file")
+	}
+
+	app := api.NewApp()
 
 	log.Print("Starting server on port 8080")
-	// handlers.CORS(handlers.AllowedHeaders([]string{"X-Requested-With", "Content-Type", "Authorization"}), handlers.AllowedMethods([]string{"GET", "POST", "PUT", "HEAD", "OPTIONS"}), handlers.AllowedOrigins([]string{"*"}))(r)
 
-	log.Fatal(http.ListenAndServe(":8080", handlers.CORS(handlers.AllowedHeaders([]string{"X-Requested-With", "Content-Type", "Authorization"}), handlers.AllowedMethods([]string{"GET", "POST", "PUT", "HEAD", "OPTIONS"}), handlers.AllowedOrigins([]string{"*"}))(r)))
-	// log.Fatal(http.ListenAndServe(":8080", r))
+	app.Run()
 }
